Moved StateResolvers along with their definitions
