+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/state_resolvers.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="sylius.order_processing.state_resolver.payment" class="Sylius\Component\Core\StateResolver\OrderPaymentStateResolver">
+            <argument type="service" id="sm.factory" />
+            <tag name="sylius.order.state_resolver" />
+        </service>
+        <service id="sylius.order_processing.state_resolver.shipment" class="Sylius\Component\Core\StateResolver\OrderShipmentStateResolver">
+            <argument type="service" id="sm.factory" />
+            <tag name="sylius.order.state_resolver" />
+        </service>
+        <service id="sylius.order_processing.state_resolver.order" class="Sylius\Component\Core\StateResolver\OrderStateResolver">
+            <argument type="service" id="sm.factory" />
+            <tag name="sylius.order.state_resolver" />
+        </service>
+    </services>
+</container>
