+++ src/Sylius/Bundle/OrderBundle/Resources/config/state_resolver.xml
@@ -13,6 +13,6 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="sylius.order_processing.order_processor" class="Sylius\Component\Order\Processor\CompositeOrderProcessor"/>
+        <service id="sylius.order_processing.state_resolver" class="Sylius\Component\Order\StateResolver\CompositeOrderStateResolver" />
     </services>
 </container>
